Reducing the HES data to alcohol/tobacco-related episodes and admissions

This is the main function used to convert the cleaned data in to summary tables of either episodes, admissions or person-specific admissions. It reads in the cleaned Hospital episode statistics for England, reducing it to the sample selection required. It assigns an attributable fraction to each row/episode based on either the broad or narrow method, for the substance chosen (alcohol or tobacco). The code then goes on to calculate a summary table of alcohol/tobacco attributable hospital episodes, admissions or person-specific admissions as chosen.

assign_risk(
  data,
  year,
  pop_data,
  youngest_age = 16,
  age_cat = c("16-17", "18-24", "25-34", "35-49", "50-64", "65-74", "75-89"),
  age_cat_start = c(16, 18, 25, 35, 50, 65, 75),
  method = c("Broad", "Narrow"),
  saf_lkup =
    read.csv("//tsclient/X/ScHARR/PR_Disease_Risk_TA/Code/Attributable_fractions/Smoking_attributable_fractions/Output/Archive/safs_long.csv"),
  aaf_lkup =
    read.csv("//tsclient/X/ScHARR/PR_Disease_Risk_TA/Code/Attributable_fractions/Alcohol_attributable_fractions/Output/Archive/aafs_long_2016.csv"),
  substance = c("Alcohol", "Tobacco"),
  level = c("Episode", "Admission"),
  summary = TRUE,
  dir = "D:/HES/working_data_files"
)

Arguments

data

Data.table - cleaned HES data.

year

Character string - indicating the year of HES data - in the form "1718".

pop_data

Data.table - mid-year population sizes.

youngest_age

Integer - Youngest age required in HES data

age_cat

Character vector of age categories. Default is c("16-17", "18-24", "25-34", "35-49", "50-64", "65-74", "75-89").

age_cat_start

Numerical vector of start ages for age categories. Default is c(16, 18, 25, 35, 50, 65, 75).

method

method required to allocate episode to condition. "Broad" or "Narrow"

saf_lkup

smoking attributable fraction long look up table in the format: "SAF", "ICD10_lookups", "disease_name", "Sex", "Age_cat". Default is for age categories: c("16-17", "18-24", "25-34", "35-49", "50-64", "65-74", "75-89"), but if summarising the HES data by other age categories, can change the attributable fraction lookup table here.

aaf_lkup

alcohol attributable fraction long look up table in the format: "Sex", "af", "ICD10_lookups", "condition", "age_cat". Default is for age categories: c("16-17", "18-24", "25-34", "35-49", "50-64", "65-74", "75-89"), but if summarising the HES data by other age categories, can change the attributable fraction lookup table here.

substance

risk. "Alcohol" or "Tobacco"

level

the level of detail of summary table. "Episode", or "Admission".

summary

Logical - whether to return a summary table or the full HES data (want this option for the costing). Defaults to TRUE. If TRUE, will output a summary table of HES data. If FALSE, then will output the whole year of data.

dir

The directory to save the output. Default is "D:/HES/working_data_files", but if summary = TRUE, can save this to the project repo.

include_youth

Logical - whether to include under 16s for youth drinking analysis. Defaults to FALSE. If TRUE, will include 11-15 as an age category, but will give them the same attributable fraction as 16-17s. If FALSE, will not include 11-15s.

Value

Returns a summary table of number of episodes/admissions/people, SAF, and attributable episodes/admissions/people, by age category, sex, imd_quintile, and condition. If summary = FALSE, will return the whole data unsummarised (for use in costing at episode level)

Examples

if (FALSE) {

assign_risk(hes, "1617", method = "Narrow", substance = "Alcohol", level = "Episode", summary = TRUE)

}