Read the HES data files

Reads in Hospital Episode Statistics for England data files for each year from 2002/2003.

read_hes(k.year.ind, test = FALSE)

Arguments

k.year.ind
is the year of HES data. In the format "yy/yy".

test
Logical - whether to run a small sample of HES data Defaults to FALSE. If FALSE, will output 1,000 rows of HES data. If TRUE, then will output the whole year of data.

Value

Returns a data table. Note that:

All variable names are converted to lower case.

Details

The variable names change between 13/14 and 14/15, and again between 16/17 and 17/18. This code reads in the HES data for one year whilst selecting only the variables we need and formatting to be the same for each year.

Variables included:

startage: age at start and end of episode. This is a derived field, calculated from episode start date (epistart) and date of birth (dob). Contains the patient's age in whole years. (Ages from 1 to 115 (1990-91 to 1994-95) and from 1 to 120 (1995-96 onwards)).

For patients under 1 year old, special codes in the range 7001 to 7007 apply:

7001 = Less than 1 day

7002 = 1 to 6 days

7003 = 7 to 28 days

7004 = 29 to 90 days (under 3 months)

7005 = 91 to 181 days (approximately 3 months to under 6 months)

7006 = 182 to 272 days (approximately 6 months to under 9 months)

7007 = 273 to 364 days (approximately 9 months to under 1 year)

null = Not applicable (other maternity event or not known)

mydob: data of birth (month and year)

dob_cfl: date of birth check flag - patient

ethnos: Ethnic category

encrypted_hesid: 1415 pseudo_hesid

admidate: contains the date the patient was admitted to hospital at the start of a hospital spell. Admidate is recorded on all episodes within a spell.

2012/13 onwards:

01/01/1800 - null date submitted

01/01/1801 - invalid date submitted

1989/90 to 2011/12:

01/01/1600 ? null date submitted

15/10/1582 ? invalid date submitted

admimeth: contains a code which identifies how the patient was admitted to hospital. Admimeth is recorded on the first and also all subsequent episodes within the spell (ie where the spell is made up of more than one episode).

Elective Admission, when the decision to admit could be separated in time from the actual admission:

11 = Waiting list

12 = Booked

13 = Planned Emergency Admission,

Non-elective (emergency admission): when admission is unpredictable and at short notice because of clinical need:

21 = Accident and emergency or dental casualty department of the Health Care Provider

22 = General Practitioner: after a request for immediate admission has been made direct to a Hospital Provider, i.e. not through a Bed bureau, by a General Practitioner: or deputy

23 = Bed bureau

24 = Consultant Clinic, of this or another Health Care Provider

25 = Admission via Mental Health Crisis Resolution Team

2A = Accident and Emergency Department of another provider where the PATIENT had not been admitted

2B = Transfer of an admitted patient from another Hospital Provider in an emergency

2C = Baby born at home as intended

2D = Other emergency admission

28 = Other means, examples are:

admitted from the Accident and Emergency Department of another provider where they had not been admitted

transfer of an admitted patient from another Hospital Provider in an emergency

baby born at home as intended Maternity Admission,

Maternity admission, of a pregnant or recently pregnant woman to a maternity ward (including delivery facilities) except when the intention is to terminate the pregnancy:

31 = Admitted ante-partum

32 = Admitted post-partum

Other admission not specified above:

82 = The birth of a baby in this Health Care Provider

83 = Baby born outside the Health Care Provider except when born at home as intended.

81 = Transfer of any admitted patient from other Hospital Provider other than in an emergency

98 = Not applicable

99 = Not known: a validation error

disdate: contains the date on which the patient was discharged from hospital. It is only present in the record for the last episode of a spell.

spelbgin: This is a derived field that contains a code that defines whether the episode is the first of a spell and whether the spell started in the current or previous year.

0 = Not first episode of spell;

1 = First episode of spell that started in previous year;

2 = First episode of spell that started in current year;

null = Not applicable

spelend: a code which defines whether the episode is the last of a spell. It is set for finished episodes (episode status - epistat - is 3) for general, delivery or birth episodes (episode type - epitype - is 1, 2 or 3) provided the discharge method (dismeth) confirms that the spell has finished. Y = Last episode of spell ; N = Not last episode of spell

speldur: This is a derived field that contains the difference in days between the admission date (admidate) and the discharge date (epiend) provided the discharge method (dismeth) confirms that the spell has finished. If the episode has not finished it is calculated from the end of the year and admidate.

5n = Duration of spell in days from 0 to 29,200;

null = Not applicable: patient not discharged (dismeth not in range 1-5), other maternity event (epitype is 5 or 6) or not valid

epiend: contains the date on which a patient left the care of a particular consultant, for one of the following reasons: Discharged from hospital (includes transfers) or moved to the care of another consultant. A null entry either indicates that the episode was unfinished at the end of the data year, or the date was unknown.

2012/13 onwards:

01/01/1800 - null date submitted

01/01/1801 - invalid date submitted

1989/90 to 2011/12:

01/01/1600

epistart: contains the date on which a patient was under the care of a particular consultant. If a patient has more than one episode in a spell, for each new episode there is a new value of epistart. However, the admission date which is copied to each new episode in a spell will remain unchanged and will be equal to the episode start date of the first episode in hospital.

2012/13 onwards:

01/01/1800 - null date submitted

01/01/1801 - invalid date submitted

1989/90 to 2011/12:

01/01/1600 ? null date submitted

15/10/1582 ? invalid date submitted

epiorder: contains the number of the episode within the current spell. All spells start with an episode where epiorder is 01. Many spells finish with this episode, but if the patient moves to the care of another consultant, a new episode begins. Episode numbers increase by 1 for each new episode until the patient is discharged (this includes transfers to another NHS trust or primary care trust - ie the first episode in the new trust will have epiorder 01). If the same patient returns for a different spell in hospital, epiorder is again set to 01. Admissions are calculated by counting the number of times epiorder is 01. When studying long term care, remember that it is not unusual to transfer psychiatric patients from one hospital to another.

2n = The number of the episode in the sequence of episodes from 01-87;

98 = Not applicable;

99 = Not known;

null = Not applicable: other maternity event

epistat: tells you whether the episode had finished before the end of the HES data-year (ie whether the episode was still 'live' at midnight on 31 March). For example, if a patient was admitted on 25 March 2005 and was not discharged (or transferred to the care of another consultant) until 4 April 2005, there will be a record describing the unfinished episode (episode status = 1) in the 2004-05 data, and a separate record describing the finished episode (episode status = 3) in the 2005-06 data. Because hospital providers are advised not to include clinical data (diagnosis and operation codes) in unfinished records, these are normally excluded from analyses. Also, if unfinished episodes are included in time series analyses - where data for more than one year is involved - there is a danger of counting the same episode twice.

classpat: identifies day cases, ordinary admissions, regular day and regular night attenders, and the special case of mothers and babies using only delivery facilities. Data about regular day and regular night attenders are not available for analysis prior to 2002-03. Since the introduction of the NHS wide clearing service in April 1996, this field has been derived from related items in the Commissioning Data Set (eg intended management).

1 = Ordinary admission

2 = Day case admission

3 = Regular day attender

4 = Regular night attender

5 = Mothers and babies using only delivery facilities

8 = Not applicable (other maternity event)

9 = Not known

mainspef: contains a code that defines the specialty under which the consultant is contracted. It can be compared with tretspef, the specialty under which the consultant worked.

tretspef: speciality of consultant

diag_xx: ICD10 coded diagnosis codes. 1-20

cause: copy of the first diagnosis code that represents an external cause

resgor: government office region of residence

resladst: local authority district of residence (old version of ONS coding system)

imd04_decile: uses IMD version 2004 on activity up to and including 2006-07; IMD version 2007 on activity between 2007-08 and 2009-10; IMD version 2010 on activity from 2010-11 to present.

epikey: record identifier

opertn_xx: Procedure codes - before april 2007 there are 12 operate codes. after there are 24.

posopdur: Number of days between the main operation and the end of the episode from 0-365. null = Not applicable: no operation or episode unfinished

Additional variables added for use in HRG code grouper:

procodet

admisorc

disdest

dismeth

epidur

PROVSPNOPS

procode3

Examples

if (FALSE) {

hes_2002 <- read_hes("0203", test = TRUE)

}