Read the Scottish Hospital Episode Statistics data files

Reads in Hospital Episode Statistics for Scotland data files for each year from 2007/08 to 2021/22. The data were provided with only the variables and disease codes needed for computing the rates of diseases related to tobacco and/or alcohol and the associated rates and costs of hospital admissions.

read_hes_scot(k.year.ind)

Arguments

k.year.ind

Integer - the year of Scottish HES data. In the format "2022".

Value

Returns a data table. Note that:

All variable names are converted to lower case.

Details

This code reads in the data for one year and produces a cleaned data file ready for further analysis.

Variables included

PATIENT_ID

AGE_IN_YEARS

SEX - 1 = male, 2 = female. 0 or 9 are not known and not specified.

ADMIDATE

ADMISSION_TYPE - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=A&ID=58&Title=Admission

ADMISSION_TRANSFER_FROM- `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=A&ID=59&Title=Admission/Transfer

DISDATE

DISCHARGE_TRANSFER_TO - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=D&ID=228&Title=Discharge/Transfer

DISCHARGE_TYPE - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=D&ID=224&Title=Discharge

SPELBGIN - 0 = Not first episode of spell; 1 = First episode of spell that started in previous year; 2 = First episode of spell that started in current year

DISCHARGE_DATE

ADMISSION_DATE

SPELDUR - the difference in days between the admission date (admidate) and the discharge date (epiend) provided the discharge method (dismeth) confirms that the spell has finished. If the episode has not finished it is calculated from the end of the year and admidate.

SPELEND - whether the episode is the last of a spell. It is set for finished episodes (episode status - epistat - is 3) for general, delivery or birth episodes (episode type - epitype - is 1, 2 or 3) provided the discharge method (dismeth) confirms that the spell has finished. Y = Last episode of spell ; N = Not last episode of spell

LENGTH_OF_STAY - The patients length of stay derived from admission date and discharge date.

EPISODE_MARKER - indicating the order of episodes of care

GLS_CIS_MARKER - SMR01 Continuous Inpatient Stay Marker (including GLS records)

MAIN_CONDITION - `https://www.ndc.scot.nhs.uk/Data-Dictionary/SMR-Datasets/General-Clinical-Information/Diagnostic-Section/Main-Condition.asp`

OTHER_CONDITION_1 - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=O&ID=365&Title=Other

OTHER_CONDITION_2

OTHER_CONDITION_3

OTHER_CONDITION_4

OTHER_CONDITION_5

ADMISSION_REASON - `https://www.ndc.scot.nhs.uk/Data-Dictionary/SMR-Datasets/Episode-Management/Admission-Reason/`

MAIN_OPERATION - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=M&ID=316&Title=Main

OTHER_OPERATION_1 - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=O&ID=368&Title=Other

OTHER_OPERATION_2

OTHER_OPERATION_3

MANAGEMENT_OF_PATIENT - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=O&ID=368&Title=Other

SPECIALTY - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=S&ID=473&Title=Specialty/Discipline`

HRG42 - Healthcare Resource Group (HRG) codes are standard groupings of clinically similar treatments which use comparable levels of healthcare resource. The field contains version 4 HRG codes.

PROVIDER_CODE - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=P&ID=932&Title=Provider

simd2016_sc_quintile - Scottish Index of Multiple Deprivation quintile of patient residence (2016 version). Least deprived = 5, most deprived = 1

EPISODE_RECORD_KEY - sequential number generated by PAS identifying the episode

Examples

if (FALSE) {

hes_2002 <- read_hes("0203", test = TRUE)

}